# public/openapi.yaml
openapi: 3.1.0
info:
  title: DBackup API
  version: 3.2.0
  description: |
    REST API for DBackup - a self-hosted database backup automation platform with encryption, compression, and smart retention.

    ## Authentication

    DBackup supports two authentication methods:

    - **Session Authentication** - Used automatically when logged in via the web UI. Session cookies are sent with each request.
    - **API Key Authentication** - For scripts, CI/CD pipelines, and external integrations. Create an API key under **Access Management → API Keys**.

    > API keys do not inherit SuperAdmin privileges. Only explicitly assigned permissions are available.

    ## Rate Limits

    | Request Type | Default Limit |
    |---|---|
    | Authentication (login, etc.) | 5/min per IP |
    | GET requests | 100/min per IP |
    | POST / PUT / DELETE | 20/min per IP |

    Rate limits are configurable in Settings → Rate Limits.
  license:
    name: GPL-3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
...
servers:
  - url: "{protocol}://{host}/api"
    description: DBackup Instance
    variables:
      protocol:
        default: https
        enum: [http, https]
      host:
        default: your-dbackup-instance.com
security:
  - BearerAuth: []
  - SessionAuth: []
...
paths:
...
  /jobs/{id}:
    put:
      tags: [Jobs]
      summary: Update Job
      operationId: updateJob
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateJobRequest"
      responses:
        "200":
          description: Job updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "409":
          description: Another job with this name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      tags: [Jobs]
      summary: Delete Job
      operationId: deleteJob
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Job deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /adapters:
    get:
      tags: [Adapters]
      summary: List Adapters
      description: |
        List all adapter configurations of a given type. Config fields are automatically decrypted in the response.

        **Permission depends on type:**
        - `database` → `sources:read`
        - `storage` → `destinations:read`
        - `notification` → `notifications:read`
      operationId: listAdapters
      security:
        - BearerAuth: []
      parameters:
        - name: type
          in: query
          required: true
          schema:
            type: string
            enum: [database, storage, notification]
          description: Adapter category
        - name: role
          in: query
          required: false
          schema:
            type: string
            enum: [DESTINATION, SOURCE]
          description: >-
            Narrows storage adapters to one role (case-insensitive). Omit to receive both.
            Ignored for other adapter types.
      responses:
        "200":
          description: Adapter list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdapterConfig"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Adapters]
      summary: Create Adapter
      description: |
        Create a new adapter configuration. Sensitive config fields (passwords, keys) are automatically encrypted before storage.

        **Permission depends on type:**
        - `database` → `sources:write`
        - `storage` → `destinations:write`
        - `notification` → `notifications:write`
      operationId: createAdapter
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAdapterRequest"
      responses:
        "201":
          description: Adapter created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdapterConfig"
        "409":
          description: Adapter with this name and type already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /adapters/test-connection:
    post:
      tags: [Connection Testing]
      summary: Test Connection
      description: Tests connectivity for a given adapter configuration. If `configId` is provided and the test succeeds, the adapter metadata is updated.
      operationId: testConnection
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [adapterId, config]
              properties:
                adapterId:
                  type: string
                  description: Adapter type identifier
                  example: mysql
                config:
                  type: object
                  description: Adapter-specific connection configuration
                configId:
                  type: string
                  description: Existing adapter ID (updates metadata on success)
                primaryCredentialId:
                  type: string
                  description: Credential profile ID to overlay on the primary connection config (username, password, access key, token, etc.)
                sshCredentialId:
                  type: string
                  description: Credential profile ID to overlay on the SSH tunnel config (SSH key or password)
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "200":
          description: Connection result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                    example: Connection successful
                  version:
                    type: string
                    example: "8.0.35"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        API Key authentication. Create an API key under **Access Management → API Keys**.

        Format: `dbackup_xxxxxxxx`
    SessionAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
      description: Session cookie (browser login)
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Resource ID
  responses:
    Unauthorized:
      description: Missing, invalid, disabled, or expired credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Valid credentials but insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Human-readable error message"
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
...
    Job:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          example: "Daily MySQL Backup"
        schedule:
          type: string
          example: "0 2 * * *"
        enabled:
          type: boolean
        sourceId:
          type: string
        compression:
          type: string
          enum: [NONE, GZIP, BROTLI]
        encryptionProfileId:
          type: [string, "null"]
        databases:
          type: string
          description: JSON array of database names
          example: '["mydb", "analytics"]'
        destinations:
          type: array
          items:
            $ref: "#/components/schemas/JobDestination"
        source:
          type: object
          properties:
            name:
              type: string
            type:
              type: string
        createdAt:
          type: string
          format: date-time
    JobDestination:
      type: object
      properties:
        id:
          type: string
        configId:
          type: string
        priority:
          type: integer
        retention:
          type: string
          description: JSON retention policy
        config:
          type: object
          properties:
            name:
              type: string
            adapterId:
              type: string
...
    UpdateJobRequest:
      type: object
      properties:
        name:
          type: string
        schedule:
          type: string
        sourceId:
          type: string
        destinations:
          type: array
          items:
            type: object
            required: [configId]
            properties:
              configId:
                type: string
              priority:
                type: integer
              retention:
                type: object
        enabled:
          type: boolean
        compression:
          type: string
          enum: [NONE, GZIP, BROTLI]
        encryptionProfileId:
          type: [string, "null"]
        notificationIds:
          type: array
          items:
            type: string
        notificationEvents:
          type: object
        databases:
          type: array
          items:
            type: string

    # ─── Executions ──────────────────────────────────────────────
...
    AdapterConfig:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          example: "Production MySQL"
        type:
          type: string
          enum: [database, storage, notification]
        adapterId:
          type: string
          description: Adapter type identifier
          example: mysql
        config:
          type: object
          description: Adapter-specific configuration (decrypted)
        metadata:
          type: [object, "null"]
          description: "Auto-detected metadata (version, status, etc.)"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        storageRole:
          type: string
          enum: [DESTINATION, SOURCE]
          default: DESTINATION
          description: >-
            Storage adapters only. A config either receives backups (DESTINATION) or
            provides folders to back up (SOURCE) - never both, because a destination writes
            job and chain folders into its configured path while a source reads folders out
            of it. Ignored for database and notification adapters.
    CreateAdapterRequest:
      type: object
      required: [name, type, adapterId, config]
      properties:
        name:
          type: string
          example: "Production MySQL"
        type:
          type: string
          enum: [database, storage, notification]
        adapterId:
          type: string
          description: |
            Adapter type identifier. Available values:

            **Database:** `mysql`, `mariadb`, `postgres`, `mongodb`, `sqlite`, `mssql`, `redis`

            **Storage:** `local-filesystem`, `s3-aws`, `s3-generic`, `s3-r2`, `s3-hetzner`, `google-drive`, `dropbox`, `onedrive`, `sftp`, `ftp`, `webdav`, `smb`, `rsync`

            **Notification:** `discord`, `slack`, `teams`, `generic-webhook`, `gotify`, `ntfy`, `telegram`, `twilio-sms`, `email`
        config:
          type: object
          description: Adapter-specific configuration. See adapter documentation for fields.
        storageRole:
          type: string
          enum: [DESTINATION, SOURCE]
          default: DESTINATION
          description: Storage adapters only - see AdapterConfig.storageRole. Ignored for other types.
...